app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
boto3>=1.34.0
botocore>=1.34.0
cryptography>=41.0.0
pyopenssl>=23.0.0
requests>=2.31.0
